Add option to generate single structure file
Instead of generating multiple files, stack the sections next to each other ina single structure.

For flat maps, they are stacked in their natural position.

For 3d and staircase maps, they are stacked on the X axis only, since an extra block is needed in the Z axis for the top row.

Feature for issue #2